Split client module
