+++ nex-rs/src/client/mod.rs
@@ -1,0 +1,7 @@
+mod connection;
+mod context;
+mod result;
+
+pub use connection::*;
+pub use context::*;
+pub use result::*;
